refactor(data): mirror src/data/ layout to src/models/
- Move data/gear/* to data/common/gear/* so the data tree mirrors
  src/models/common/gear/ (which regrouped under common in step 2).
- Extract the inline DATA arrays from abilities.ts (6 entries) and
  skills.ts (18 entries) into data/common/abilities.json and
  data/common/skills.json. The narrow id unions, types, and the model
  object stay in the .ts file; the tables are just catalog data.

src/data/ and src/models/ now share the same folder shape.
