+++ src/data/common/gear/armor.json
