+++ src/data/common/gear/weapon-properties.json
