+++ src/data/common/gear/weapons.json
