+++ src/models/common/abilities.ts
@@ -1,3 +1,5 @@
+import abilitiesData from "src/data/common/abilities.json";
+
 export type AbilityName = "str" | "dex" | "con" | "int" | "wis" | "cha";
 export type AbilityScores = Record<AbilityName, number>;
 
@@ -7,15 +9,7 @@ export type AbilityDetails = {
   short: string;
 };
 
-const DATA: AbilityDetails[] = [
-  { id: "str", label: "Strength", short: "STR" },
-  { id: "dex", label: "Dexterity", short: "DEX" },
-  { id: "con", label: "Constitution", short: "CON" },
-  { id: "int", label: "Intelligence", short: "INT" },
-  { id: "wis", label: "Wisdom", short: "WIS" },
-  { id: "cha", label: "Charisma", short: "CHA" },
-];
-
+const DATA = abilitiesData as AbilityDetails[];
 const BY_ID = new Map(DATA.map((a) => [a.id, a]));
 
 export const abilities = {
